+++ BugsnagPerformance/Assets/BugsnagPerformance/Scripts/Public/Span.cs
@@ -26,8 +26,12 @@ namespace BugsnagUnityPerformance
         internal int DroppedAttributesCount;
         private int _customAttributeCount;
         private int _maxCustomAttributes;
+        private FrameMetricsSnapshot _startFrameRateMetricsSnapshot;
 
-        public Span(string name, SpanKind kind, string id, string traceId, string parentSpanId, DateTimeOffset startTime, bool? isFirstClass, OnSpanEnd onSpanEnd, int maxCustomAttributes)
+        public Span(string name, SpanKind kind, string id, 
+        string traceId, string parentSpanId, DateTimeOffset startTime, 
+        bool? isFirstClass, OnSpanEnd onSpanEnd, int maxCustomAttributes,
+        FrameMetricsSnapshot startFrameRateMetricsSnapshot)
         {
             Name = name ?? string.Empty;
             Kind = kind;
@@ -41,6 +45,7 @@ namespace BugsnagUnityPerformance
             {
                 SetAttributeInternal("bugsnag.span.first_class", isFirstClass.Value);
             }
+            _startFrameRateMetricsSnapshot = startFrameRateMetricsSnapshot;
             _onSpanEnd = onSpanEnd;
         }
 
@@ -213,6 +218,24 @@ namespace BugsnagUnityPerformance
         internal void SetCallbackComplete()
         {
             _callbackComplete = true;
+        }
+
+        internal void CalculateFrameRateMetrics(FrameMetricsSnapshot endFrameRateMetricsSnapshot)
+        {
+            if(_startFrameRateMetricsSnapshot == null || endFrameRateMetricsSnapshot == null)
+            {
+                return;
+            }
+            SetAttributeInternal("bugsnag.framerate.frozen_frames", endFrameRateMetricsSnapshot.FrozenFrames - _startFrameRateMetricsSnapshot.FrozenFrames);
+            SetAttributeInternal("bugsnag.framerate.slow_frames", endFrameRateMetricsSnapshot.SlowFrames - _startFrameRateMetricsSnapshot.SlowFrames);
+            SetAttributeInternal("bugsnag.framerate.total_frames", endFrameRateMetricsSnapshot.TotalFrames - _startFrameRateMetricsSnapshot.TotalFrames);
+        }
+
+        internal void RemoveFrameRateMetrics()
+        {
+            _attributes.Remove("bugsnag.framerate.frozen_frames");
+            _attributes.Remove("bugsnag.framerate.slow_frames");
+            _attributes.Remove("bugsnag.framerate.total_frames");
         }
     }
 }