+++ BugsnagPerformance/Assets/UnitTests/SamplerTests.cs
@@ -52,7 +52,8 @@ namespace Tests
                 DateTimeOffset.Now,
                 true,
                 OnSpanEnd,
-                128);
+                128,
+                null);
 
             Assert.AreEqual(1.0, sampler.Probability);
             Assert.IsTrue(sampler.Sampled(span));
@@ -91,7 +92,8 @@ namespace Tests
                 DateTimeOffset.Now,
                 true,
                 OnSpanEnd,
-                128);
+                128,
+                null);
 
             Assert.IsTrue(sampler.Sampled(span));
         }
@@ -111,7 +113,8 @@ namespace Tests
                 DateTimeOffset.Now,
                 true,
                 OnSpanEnd,
-                128);
+                128,
+                null);
 
             Assert.IsFalse(sampler.Sampled(span));
         }
@@ -131,7 +134,8 @@ namespace Tests
                 DateTimeOffset.Now,
                 true,
                 OnSpanEnd,
-                128);
+                128,
+                null);
 
             Assert.IsFalse(sampler.Sampled(span));
 
@@ -143,7 +147,8 @@ namespace Tests
                 DateTimeOffset.Now,
                 true,
                 OnSpanEnd,
-                128);
+                128,
+                null);
             Assert.IsTrue(sampler.Sampled(span));
 
         }
@@ -163,7 +168,8 @@ namespace Tests
                 DateTimeOffset.Now,
                 true,
                 OnSpanEnd,
-                128);
+                128,
+                null);
 
             Assert.IsTrue(sampler.Sampled(span));
 
@@ -175,7 +181,8 @@ namespace Tests
                 DateTimeOffset.Now,
                 true,
                 OnSpanEnd,
-                128);
+                128,
+                null);
             Assert.IsTrue(sampler.Sampled(span));
 
             span = new Span("test",
@@ -186,7 +193,8 @@ namespace Tests
                 DateTimeOffset.Now,
                 true,
                 OnSpanEnd,
-                128);
+                128,
+                null);
             Assert.IsFalse(sampler.Sampled(span));
 
         }